refactor(#223): fold parse_host_port into pentest_core::config
Address review nits on PR #225:

- Extract `parse_host_port` into `ConnectorConfig::split_authority` in
  pentest-core so `normalize_host` and the pre-connect probe share one
  parser (was hand-rolled in both places, per the reviewer's
  reimplementation nit). Tests move alongside — six new
  `split_authority_*` cases join the existing normalize-host coverage.
- Simplify the `use_tls` computation in `probe_host_reachable` to a
  single `.trim().to_lowercase()` binding (was up to four calls) and
  drop the `matches!` guard-pattern in favor of a straight boolean.
- Add a comment above the reqwest builder noting `MATRIX_TLS_INSECURE`
  mirrors the existing convention in `token_refresh` and
  `pentest_core::matrix::client` — self-signed dev clusters only; the
  probe-scoped relaxation does not touch the SDK's own TLS trust.

No behavior change against StrikeHub-launched Pick: the probe still
runs the same TCP+TLS handshake, `MATRIX_AUTH_TOKEN` still short-circuits
the chat-panel auth chain first (StrikeHub already injects it), and
`STRIKE48_API_URL` stays untouched (StrikeHub sets it to the local
proxy origin; standalone leaves it unset for warn+permit).
